+++ src/components/example/ExampleAxios.vue
@@ -35,7 +35,19 @@ const renderingIsDone = () => {
 
 <template>
   <div class="border border-gray-200 p-4 rounded-lg flex flex-col items-center basis-xs">
-    <h3 class="border-b pb-1 mb-4">Axios + Tanstack Query</h3>
+    <div class="border-b border-gray-300 pb-1 mb-4 w-full text-center">
+      <h3>Axios + Tanstack Vue Query</h3>
+      <div class="leading-0">
+        <a
+          href="https://github.com/roy-margasa/YAVBT/blob/master/src/components/example/ExampleAxios.vue"
+          target="_blank"
+          rel="noopener noreferrer"
+          class="text-blue-700 text-xs"
+        >
+          source
+        </a>
+      </div>
+    </div>
     <div class="w-full flex flex-col items-center">
       <div v-if="isLoadingMedia" class="w-full animate-pulse">
         <div class="w-full block h-30 bg-gray-300"></div>