+++ src/components/example/ExampleVeeValidate.vue
@@ -45,7 +45,19 @@ const onSubmitComponent = (values: Record<string, unknown>) => {
 
 <template>
   <div class="border border-gray-200 p-4 rounded-lg flex flex-col items-center basis-xs">
-    <h3 class="border-b pb-1 mb-4">Vee Validate + yup</h3>
+    <div class="border-b border-gray-300 pb-1 mb-4 w-full text-center">
+      <h3>Vee Validate + yup</h3>
+      <div class="leading-0">
+        <a
+          href="https://github.com/roy-margasa/YAVBT/blob/master/src/components/example/ExampleVeeValidate.vue"
+          target="_blank"
+          rel="noopener noreferrer"
+          class="text-blue-700 text-xs"
+        >
+          source
+        </a>
+      </div>
+    </div>
 
     <div class="w-full text-center">
       <p>Composables</p>